requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
jinja2==3.1.2
python-multipart==0.0.6
python-markdown==3.5.1
google-cloud-firestore==2.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
starlette==0.27.0
pydantic-settings==2.0.3